Redirect Symbolic API to NNVM
