+++ include/mxnet/base.h
@@ -11,6 +11,8 @@
 #include <dmlc/type_traits.h>
 #include <dmlc/parameter.h>
 #include <mshadow/tensor.h>
+// nnvm headers for symbolic construction.
+#include <nnvm/op.h>
 #include <nnvm/tuple.h>
 #include <string>
 
@@ -83,6 +85,8 @@ typedef mshadow::index_t index_t;
 typedef mshadow::default_real_t real_t;
 /*! \brief Shape data structure used to record shape information */
 using TShape = nnvm::TShape;
+/*! \brief operator structure from NNVM */
+using Op = nnvm::Op;
 
 /*! \brief Context information about the execution enviroment */
 struct Context {
