+++ src/c_api/c_api.cc
@@ -404,502 +404,6 @@ int MXFuncInvokeEx(FunctionHandle fun,
   API_END();
 }
 
-//--------------------------------------------
-// Part 3: symbolic configuration generation
-//--------------------------------------------
-
-int MXSymbolListAtomicSymbolCreators(mx_uint *out_size,
-                                     AtomicSymbolCreator **out_array) {
-  API_BEGIN();
-  auto &vec = dmlc::Registry<OperatorPropertyReg>::List();
-  *out_size = static_cast<mx_uint>(vec.size());
-  *out_array = (AtomicSymbolCreator*)(dmlc::BeginPtr(vec));  //  NOLINT(*)
-  API_END();
-}
-
-int MXSymbolGetAtomicSymbolName(AtomicSymbolCreator creator,
-                                const char **out) {
-  API_BEGIN();
-  OperatorPropertyReg *e = static_cast<OperatorPropertyReg *>(creator);
-  *out = e->name.c_str();
-  API_END();
-}
-
-int MXSymbolGetAtomicSymbolInfo(AtomicSymbolCreator creator,
-                                const char **name,
-                                const char **description,
-                                mx_uint *num_args,
-                                const char ***arg_names,
-                                const char ***arg_type_infos,
-                                const char ***arg_descriptions,
-                                const char **key_var_num_args,
-                                const char **return_type) {
-  OperatorPropertyReg *e = static_cast<OperatorPropertyReg *>(creator);
-  *key_var_num_args = e->key_var_num_args.c_str();
-  return MXAPIGetFunctionRegInfo(e, name, description, num_args,
-                                 arg_names, arg_type_infos, arg_descriptions,
-                                 return_type);
-}
-
-int MXSymbolCreateAtomicSymbol(AtomicSymbolCreator creator,
-                               mx_uint num_param,
-                               const char **keys,
-                               const char **vals,
-                               SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  OperatorProperty *op = nullptr;
-
-  API_BEGIN();
-  OperatorPropertyReg *e = static_cast<OperatorPropertyReg *>(creator);
-  op = e->body();
-  std::vector<std::pair<std::string, std::string> > kwargs;
-  for (mx_uint i = 0; i < num_param; ++i) {
-    kwargs.push_back({std::string(keys[i]), std::string(vals[i])});
-  }
-  op->Init(kwargs);
-  *s = Symbol::Create(op);
-  *out = s;
-  API_END_HANDLE_ERROR(delete s; delete op);
-}
-
-int MXSymbolCreateVariable(const char *name, SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  API_BEGIN();
-  *s = Symbol::CreateVariable(name);
-  *out = s;
-  API_END_HANDLE_ERROR(delete s);
-}
-
-int MXSymbolCreateGroup(mx_uint num_symbols,
-                        SymbolHandle *symbols,
-                        SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  Symbol **sym_arr = (Symbol**)symbols; // NOLINT(*)
-  API_BEGIN();
-  std::vector<Symbol> syms;
-  for (mx_uint i = 0; i < num_symbols; ++i) {
-    syms.push_back(*sym_arr[i]);
-  }
-  *s = Symbol::CreateGroup(syms);
-  *out = s;
-  API_END_HANDLE_ERROR(delete s);
-}
-
-int MXSymbolGetOutput(SymbolHandle symbol,
-                      mx_uint index,
-                      SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  API_BEGIN();
-  *s = (*static_cast<Symbol*>(symbol))[index];
-  *out = s;
-  API_END_HANDLE_ERROR(delete s);
-}
-
-int MXSymbolGetInternals(SymbolHandle symbol,
-                         SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  API_BEGIN();
-  *s = static_cast<Symbol*>(symbol)->GetInternals();
-  *out = s;
-  API_END_HANDLE_ERROR(delete s);
-}
-
-int MXSymbolCreateFromFile(const char *fname, SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  API_BEGIN();
-  std::unique_ptr<dmlc::Stream> fi(dmlc::Stream::Create(fname, "r"));
-  dmlc::istream is(fi.get());
-  dmlc::JSONReader reader(&is);
-  s->Load(&reader);
-  // reset file pointer
-  is.set_stream(nullptr);
-  *out = s;
-  API_END_HANDLE_ERROR(delete s);
-}
-
-int MXSymbolCreateFromJSON(const char *json, SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  API_BEGIN();
-  std::string buf(json);
-  std::istringstream is(buf);
-  dmlc::JSONReader reader(&is);
-  s->Load(&reader);
-  *out = s;
-  API_END_HANDLE_ERROR(delete s);
-}
-
-int MXSymbolSaveToFile(SymbolHandle symbol, const char *fname) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  API_BEGIN();
-  std::unique_ptr<dmlc::Stream> fo(dmlc::Stream::Create(fname, "w"));
-  dmlc::ostream os(fo.get());
-  dmlc::JSONWriter writer(&os);
-  s->Save(&writer);
-  // reset file pointer, force flush
-  os.set_stream(nullptr);
-  API_END();
-}
-
-int MXSymbolSaveToJSON(SymbolHandle symbol, const char **out_json) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  std::ostringstream os;
-  dmlc::JSONWriter writer(&os);
-  s->Save(&writer);
-  ret->ret_str = os.str();
-  *out_json = ret->ret_str.c_str();
-  API_END();
-}
-
-int MXSymbolFree(SymbolHandle symbol) {
-  API_BEGIN();
-  delete static_cast<Symbol*>(symbol);
-  API_END();
-}
-
-int MXSymbolCopy(SymbolHandle symbol, SymbolHandle *out) {
-  Symbol *s = new Symbol();
-  API_BEGIN();
-  *s = static_cast<const Symbol*>(symbol)->Copy();
-  *out = s;
-  API_END_HANDLE_ERROR(delete s);
-}
-
-int MXSymbolPrint(SymbolHandle symbol, const char **out_str) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  std::ostringstream os;
-  s->Print(os);
-  ret->ret_str = os.str();
-  *out_str = (ret->ret_str).c_str();
-  API_END();
-}
-
-int MXSymbolGetName(SymbolHandle symbol,
-                    const char** out,
-                    int* success) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  if (s->GetName(&(ret->ret_str))) {
-    *out = (ret->ret_str).c_str();
-    *success = 1;
-  } else {
-    *out = nullptr;
-    *success = 0;
-  }
-  API_END();
-}
-
-int MXSymbolGetAttr(SymbolHandle symbol,
-                    const char* key,
-                    const char** out,
-                    int* success) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  if (s->GetAttr(key, &(ret->ret_str))) {
-    *out = (ret->ret_str).c_str();
-    *success = 1;
-  } else {
-    *out = nullptr;
-    *success = 0;
-  }
-  API_END();
-}
-
-int MXSymbolSetAttr(SymbolHandle symbol,
-                    const char* key,
-                    const char* value) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  API_BEGIN();
-  s->SetAttr(key, value);
-  API_END();
-}
-
-int _MXSymbolListAttrImpl(SymbolHandle symbol,
-                          bool shalow,
-                          mx_uint *out_size,
-                          const char*** out) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  std::map<std::string, std::string> attr =
-      std::move(shalow ? s->ListAttrShallow() : s->ListAttr());
-  std::vector<std::string> attrList;
-  *out_size = 0;
-  for (auto it : attr) {
-    attrList.push_back(it.first);
-    attrList.push_back(it.second);
-    (*out_size)++;
-  }
-
-  ret->ret_vec_str = std::move(attrList);
-  ret->ret_vec_charp.clear();
-  for (size_t i = 0; i < ret->ret_vec_str.size(); ++i) {
-    ret->ret_vec_charp.push_back(ret->ret_vec_str[i].c_str());
-  }
-  *out = dmlc::BeginPtr(ret->ret_vec_charp);
-  API_END();
-}
-
-int MXSymbolListAttr(SymbolHandle symbol,
-                     mx_uint *out_size,
-                     const char*** out) {
-  return _MXSymbolListAttrImpl(symbol, false, out_size, out);
-}
-
-int MXSymbolListAttrShallow(SymbolHandle symbol,
-                            mx_uint *out_size,
-                            const char*** out) {
-  return _MXSymbolListAttrImpl(symbol, true, out_size, out);
-}
-
-int MXSymbolListArguments(SymbolHandle symbol,
-                          mx_uint *out_size,
-                          const char ***out_str_array) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  ret->ret_vec_str = std::move(s->ListArguments());
-  ret->ret_vec_charp.clear();
-  for (size_t i = 0; i < ret->ret_vec_str.size(); ++i) {
-    ret->ret_vec_charp.push_back(ret->ret_vec_str[i].c_str());
-  }
-  *out_size = static_cast<mx_uint>(ret->ret_vec_charp.size());
-  *out_str_array = dmlc::BeginPtr(ret->ret_vec_charp);
-  API_END();
-}
-
-int MXSymbolListOutputs(SymbolHandle symbol,
-                        mx_uint *out_size,
-                        const char ***out_str_array) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  ret->ret_vec_str = std::move(s->ListOutputs());
-  ret->ret_vec_charp.clear();
-  for (size_t i = 0; i < ret->ret_vec_str.size(); ++i) {
-    ret->ret_vec_charp.push_back(ret->ret_vec_str[i].c_str());
-  }
-  *out_size = static_cast<mx_uint>(ret->ret_vec_charp.size());
-  *out_str_array = dmlc::BeginPtr(ret->ret_vec_charp);
-  API_END();
-}
-
-int MXSymbolListAuxiliaryStates(SymbolHandle symbol,
-                                mx_uint *out_size,
-                                const char ***out_str_array) {
-  Symbol *s = static_cast<Symbol*>(symbol);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  API_BEGIN();
-  ret->ret_vec_str = std::move(s->ListAuxiliaryStates());
-  ret->ret_vec_charp.clear();
-  for (size_t i = 0; i < ret->ret_vec_str.size(); ++i) {
-    ret->ret_vec_charp.push_back(ret->ret_vec_str[i].c_str());
-  }
-  *out_size = static_cast<mx_uint>(ret->ret_vec_charp.size());
-  *out_str_array = dmlc::BeginPtr(ret->ret_vec_charp);
-  API_END();
-}
-
-int MXSymbolCompose(SymbolHandle sym,
-                    const char *name,
-                    mx_uint num_args,
-                    const char** keys,
-                    SymbolHandle* args) {
-  API_BEGIN();
-  std::string s_name;
-  if (name != nullptr) s_name = name;
-
-  Symbol* s = static_cast<Symbol*>(sym);
-  if (keys == nullptr && num_args != 0) {
-    std::vector<Symbol> pos_args;
-    for (mx_uint i = 0; i < num_args; ++i) {
-      pos_args.push_back(*((Symbol*)args[i]));  //  NOLINT(*)
-    }
-    s->Compose(pos_args, s_name);
-  } else {
-    std::unordered_map<std::string, Symbol> kwargs;
-    for (mx_uint i = 0; i < num_args; ++i) {
-      kwargs[keys[i]] = *((Symbol*)args[i]);  //  NOLINT(*)
-    }
-    s->Compose(kwargs, s_name);
-  }
-  API_END();
-}
-
-int MXSymbolGrad(SymbolHandle sym, mx_uint num_wrt, const char** wrt, SymbolHandle* out) {
-  API_BEGIN();
-  Symbol* s = static_cast<Symbol*>(sym);
-  std::vector<std::string> wrts(num_wrt);
-  for (mx_uint i = 0; i < num_wrt; ++i) {
-    wrts[i] = wrt[i];
-  }
-  Symbol* ret = new Symbol;
-  *ret = s->Grad(wrts);
-  *out = ret;
-  API_END();
-}
-
-int MXSymbolInferShape(SymbolHandle sym,
-                       mx_uint num_args,
-                       const char** keys,
-                       const mx_uint *arg_ind_ptr,
-                       const mx_uint *arg_shape_data,
-                       mx_uint *in_shape_size,
-                       const mx_uint **in_shape_ndim,
-                       const mx_uint ***in_shape_data,
-                       mx_uint *out_shape_size,
-                       const mx_uint **out_shape_ndim,
-                       const mx_uint ***out_shape_data,
-                       mx_uint *aux_shape_size,
-                       const mx_uint **aux_shape_ndim,
-                       const mx_uint ***aux_shape_data,
-                       int *complete) {
-  Symbol *s = static_cast<Symbol*>(sym);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  bool succ;
-  API_BEGIN();
-  if (keys == nullptr && num_args != 0) {
-    ret->arg_shapes.clear();
-    for (mx_uint i = 0; i < num_args; ++i) {
-      ret->arg_shapes.push_back(TShape(arg_shape_data + arg_ind_ptr[i],
-                                       arg_shape_data + arg_ind_ptr[i+1]));
-    }
-    succ = s->InferShape(&(ret->arg_shapes), &(ret->out_shapes), &(ret->aux_shapes));
-  } else {
-    std::unordered_map<std::string, TShape> kwargs;
-    for (mx_uint i = 0; i < num_args; ++i) {
-      kwargs[keys[i]] = TShape(arg_shape_data + arg_ind_ptr[i],
-                               arg_shape_data + arg_ind_ptr[i+1]);
-    }
-    succ = s->InferShape(kwargs, &(ret->arg_shapes), &(ret->out_shapes), &(ret->aux_shapes));
-  }
-  if (succ) {
-    MXAPIThreadLocalEntry::SetupShapeArrayReturn(
-        ret->arg_shapes, &(ret->arg_shape_ndim), &(ret->arg_shape_data));
-    MXAPIThreadLocalEntry::SetupShapeArrayReturn(
-        ret->out_shapes, &(ret->out_shape_ndim), &(ret->out_shape_data));
-    MXAPIThreadLocalEntry::SetupShapeArrayReturn(
-        ret->aux_shapes, &(ret->aux_shape_ndim), &(ret->aux_shape_data));
-    *in_shape_size = static_cast<mx_uint>(ret->arg_shapes.size());
-    *in_shape_ndim = dmlc::BeginPtr(ret->arg_shape_ndim);
-    *in_shape_data = dmlc::BeginPtr(ret->arg_shape_data);
-    *out_shape_size = static_cast<mx_uint>(ret->out_shapes.size());
-    *out_shape_ndim = dmlc::BeginPtr(ret->out_shape_ndim);
-    *out_shape_data = dmlc::BeginPtr(ret->out_shape_data);
-    *aux_shape_size = static_cast<mx_uint>(ret->aux_shapes.size());
-    *aux_shape_ndim = dmlc::BeginPtr(ret->aux_shape_ndim);
-    *aux_shape_data = dmlc::BeginPtr(ret->aux_shape_data);
-    *complete = 1;
-  } else {
-    *complete = 0;
-  }
-  API_END();
-}
-
-int MXSymbolInferShapePartial(SymbolHandle sym,
-                       mx_uint num_args,
-                       const char** keys,
-                       const mx_uint *arg_ind_ptr,
-                       const mx_uint *arg_shape_data,
-                       mx_uint *in_shape_size,
-                       const mx_uint **in_shape_ndim,
-                       const mx_uint ***in_shape_data,
-                       mx_uint *out_shape_size,
-                       const mx_uint **out_shape_ndim,
-                       const mx_uint ***out_shape_data,
-                       mx_uint *aux_shape_size,
-                       const mx_uint **aux_shape_ndim,
-                       const mx_uint ***aux_shape_data,
-                       int *complete) {
-  Symbol *s = static_cast<Symbol*>(sym);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  bool succ;
-  API_BEGIN();
-  if (keys == nullptr && num_args != 0) {
-    ret->arg_shapes.clear();
-    for (mx_uint i = 0; i < num_args; ++i) {
-      ret->arg_shapes.push_back(TShape(arg_shape_data + arg_ind_ptr[i],
-                                       arg_shape_data + arg_ind_ptr[i+1]));
-    }
-    succ = s->InferShape(&(ret->arg_shapes), &(ret->out_shapes), &(ret->aux_shapes), true);
-  } else {
-    std::unordered_map<std::string, TShape> kwargs;
-    for (mx_uint i = 0; i < num_args; ++i) {
-      kwargs[keys[i]] = TShape(arg_shape_data + arg_ind_ptr[i],
-                               arg_shape_data + arg_ind_ptr[i+1]);
-    }
-    succ = s->InferShape(kwargs, &(ret->arg_shapes), &(ret->out_shapes), &(ret->aux_shapes), true);
-  }
-  if (succ) {
-    MXAPIThreadLocalEntry::SetupShapeArrayReturn(
-        ret->arg_shapes, &(ret->arg_shape_ndim), &(ret->arg_shape_data));
-    MXAPIThreadLocalEntry::SetupShapeArrayReturn(
-        ret->out_shapes, &(ret->out_shape_ndim), &(ret->out_shape_data));
-    MXAPIThreadLocalEntry::SetupShapeArrayReturn(
-        ret->aux_shapes, &(ret->aux_shape_ndim), &(ret->aux_shape_data));
-    *in_shape_size = static_cast<mx_uint>(ret->arg_shapes.size());
-    *in_shape_ndim = dmlc::BeginPtr(ret->arg_shape_ndim);
-    *in_shape_data = dmlc::BeginPtr(ret->arg_shape_data);
-    *out_shape_size = static_cast<mx_uint>(ret->out_shapes.size());
-    *out_shape_ndim = dmlc::BeginPtr(ret->out_shape_ndim);
-    *out_shape_data = dmlc::BeginPtr(ret->out_shape_data);
-    *aux_shape_size = static_cast<mx_uint>(ret->aux_shapes.size());
-    *aux_shape_ndim = dmlc::BeginPtr(ret->aux_shape_ndim);
-    *aux_shape_data = dmlc::BeginPtr(ret->aux_shape_data);
-    *complete = 1;
-  } else {
-    *complete = 0;
-  }
-  API_END();
-}
-
-int MXSymbolInferType(SymbolHandle sym,
-                      mx_uint num_args,
-                      const char** keys,
-                      const int *arg_type_data,
-                      mx_uint *in_type_size,
-                      const int **in_type_data,
-                      mx_uint *out_type_size,
-                      const int **out_type_data,
-                      mx_uint *aux_type_size,
-                      const int **aux_type_data,
-                      int *complete) {
-  Symbol *s = static_cast<Symbol*>(sym);
-  MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
-  bool succ;
-  API_BEGIN();
-  if (keys == nullptr && num_args != 0) {
-    ret->arg_types.clear();
-    for (mx_uint i = 0; i < num_args; ++i) {
-      ret->arg_types.push_back(arg_type_data[i]);
-    }
-    succ = s->InferType(&(ret->arg_types), &(ret->out_types), &(ret->aux_types));
-  } else {
-    std::unordered_map<std::string, int> kwargs;
-    for (mx_uint i = 0; i < num_args; ++i) {
-      kwargs[keys[i]] = arg_type_data[i];
-    }
-    succ = s->InferType(kwargs, &(ret->arg_types), &(ret->out_types), &(ret->aux_types));
-  }
-  if (succ) {
-    *in_type_size = static_cast<mx_uint>(ret->arg_types.size());
-    *in_type_data = dmlc::BeginPtr(ret->arg_types);
-    *out_type_size = static_cast<mx_uint>(ret->out_types.size());
-    *out_type_data = dmlc::BeginPtr(ret->out_types);
-    *aux_type_size = static_cast<mx_uint>(ret->aux_types.size());
-    *aux_type_data = dmlc::BeginPtr(ret->aux_types);
-    *complete = 1;
-  } else {
-    *complete = 0;
-  }
-  API_END();
-}
-
 int MXExecutorPrint(ExecutorHandle handle, const char **out_str) {
   Executor *exec = static_cast<Executor*>(handle);
   MXAPIThreadLocalEntry *ret = MXAPIThreadLocalStore::Get();
