+++ src/c_api/c_api_error.cc
@@ -3,18 +3,13 @@
  * \file c_api_error.cc
  * \brief C error handling
  */
+#include <nnvm/c_api.h>
 #include "./c_api_common.h"
 
-struct ErrorEntry {
-  std::string last_error;
-};
-
-typedef dmlc::ThreadLocalStore<ErrorEntry> MXAPIErrorStore;
-
 const char *MXGetLastError() {
-  return MXAPIErrorStore::Get()->last_error.c_str();
+  return NNGetLastError();
 }
 
 void MXAPISetLastError(const char* msg) {
-  MXAPIErrorStore::Get()->last_error = msg;
+  NNAPISetLastError(msg);
 }
